Makes Searchspace a sorted iterable (#39)
This PR adds functionality to use the `Searchspace` object as a sorted iterable, similar to Python standard dictionaries.

Three new methods are added to the `Searchspace` class:
`.items()`
`.keys()`
`.values()`

All of these return iterable objects that return the hyperparameters in the same order, i.e. the order in which they were added to the searchspace by the user at the time of definition.